requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
isodate==0.6.1
lxml==4.9.2
platformdirs==3.2.0
PyJWT==2.6.0
pyodbc==4.0.39
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==0.10.1
simple-salesforce==1.12.3
six==1.16.0
urllib3==1.26.15
zeep==4.2.1
